refactor(macos): extract shell settings views
