+++ clients/apple/README.md
@@ -23,8 +23,8 @@ recorded in [`ARCHITECTURE.md`](./ARCHITECTURE.md).
   placement support, plus small AppKit-backed shell support adapters
 - `MacShellRootView.swift`: thin primary macOS shell composition root
 - `Views/Shell/`: shell composition plus focused sidebar, Space slider, tab-row,
-  drop-target, activity, pane-topology, workspace, and command-palette SwiftUI
-  components
+  drop-target, activity, pane-topology, workspace, command-palette, and settings
+  SwiftUI components
 - `Models/Shell/`: focused shell, Terminal Profile, managed-account, activity,
   context, and pane/content/tree/tab/Space/workspace snapshot families; runtime support,
   focused presentation models, and settings navigation, Terminal Profile, managed-user,
@@ -37,7 +37,8 @@ recorded in [`ARCHITECTURE.md`](./ARCHITECTURE.md).
   keep runtime metadata and IO out of the observable host
 - `Services/Terminal/`: terminal host runtime reporting, window observation,
   agent-activity projection, and terminal runtime service collaborators
-- `TerminalPaneView.swift` / `TerminalHostView.swift`: current terminal pane and host surfaces
+- `TerminalPaneView.swift` / `TerminalHostView.swift`: current terminal pane and host surfaces;
+  settings presentation lives under `Views/Shell/Settings/`
 - `ShellHostController.swift`: observable shell state and root controller lifecycle
 - `ShellControlPlane.swift`: thin shell control-plane orchestration across the
   shell service owners
