+++ clients/apple/alan-macos/TerminalPaneView.swift
@@ -1,5 +1,4 @@
 import Foundation
-import OSLog
 import SwiftUI
 
 struct TerminalPaneView: View {
@@ -1709,1331 +1708,6 @@ private enum ShellMarkdownContentLoader {
 private enum ShellMarkdownContentLoadResult {
     case success(AttributedString)
     case failure
-}
-
-private struct ShellSettingsContentView: View {
-    let descriptor: ShellContentRenderDescriptor
-
-    @AppStorage("alanShellAppearanceMode") private var appearanceMode = ShellAppearanceMode.system
-    @AppStorage("alanShellSidebarCollapsed") private var isSidebarCollapsed = false
-    @AppStorage("alanShellDimsInactiveSplitPanes") private var dimsInactiveSplitPanes = true
-    @AppStorage(AlanPerformanceDiagnosticsController.preferenceKey)
-    private var performanceDiagnosticsEnabled = false
-    @State private var localSummary = ShellSettingsLocalSummary.current()
-    @State private var terminalProfilesSummary = TerminalProfileSettingsSummary.current()
-    @State private var privilegedHelperSummary = PrivilegedHelperSettingsSummary.current()
-    @State private var managedTerminalAccountsSummary = ManagedTerminalAccountSettingsSummary.empty
-    @State private var lastDiagnosticsExportURL: URL?
-    @State private var selectedGroup = ShellSettingsNavigationGroup.general
-    @State private var isManagedUserCreationPresented = false
-    @State private var managedUserCreationDraft = ManagedTerminalUserCreationDraft(
-        unixUserName: "",
-        displayLabel: ""
-    )
-    @State private var managedUserCreationPreviewResult: ManagedTerminalUserCreationPreviewResult?
-    @State private var managedUserActionSheet: ShellManagedUserActionSheetState?
-    @State private var managedUserApplyDiagnostics: [String] = []
-    @State private var managedUserApplyInFlight = false
-
-    nonisolated private static let managedUserApplyLogger = Logger(
-        subsystem: Bundle.main.bundleIdentifier ?? "app.alanworks.macos",
-        category: "ManagedUsers"
-    )
-    nonisolated private static let managedUserApplyTimeoutNanoseconds: UInt64 =
-        10 * 60 * 1_000_000_000
-
-    private var snapshot: ShellSettingsSurfaceSnapshot {
-        ShellSettingsSurfaceSnapshot.make(
-            local: localSummary,
-            terminalProfiles: terminalProfilesSummary,
-            privilegedHelper: privilegedHelperSummary,
-            managedTerminalAccounts: managedTerminalAccountsSummary,
-            diagnostics: diagnosticsSummary
-        )
-    }
-
-    private var settingsGroups: [ShellSettingsNavigationGroupModel] {
-        snapshot.navigationGroups
-    }
-
-    private var selectedGroupModel: ShellSettingsNavigationGroupModel {
-        settingsGroups.first { $0.id == selectedGroup }
-            ?? settingsGroups.first
-            ?? ShellSettingsNavigationGroupModel(id: .general, sections: [])
-    }
-
-    var body: some View {
-        ZStack {
-            ShellSettingsBackdrop()
-
-            HStack(alignment: .top, spacing: 0) {
-                ShellSettingsNavigationView(
-                    groups: settingsGroups,
-                    selectedGroup: $selectedGroup
-                )
-                .frame(width: ShellSettingsMetrics.navigationWidth, alignment: .topLeading)
-                .padding(.leading, ShellSettingsMetrics.navigationLeadingPadding)
-                .padding(.trailing, ShellSettingsMetrics.navigationTrailingPadding)
-                .padding(.top, ShellSettingsMetrics.navigationTopPadding)
-                .frame(maxHeight: .infinity, alignment: .topLeading)
-                .background {
-                    ShellSettingsNavigationRailBackground()
-                }
-
-                ZStack(alignment: .topLeading) {
-                    ShellSettingsDetailBackground()
-
-                    ScrollView {
-                        ShellSettingsGroupView(
-                            group: selectedGroupModel,
-                            appearanceMode: $appearanceMode,
-                            sidebarVisible: sidebarVisible,
-                            dimsInactiveSplitPanes: $dimsInactiveSplitPanes,
-                            performanceDiagnosticsEnabled: performanceDiagnosticsBinding,
-                            onExportPerformanceDiagnostics: exportPerformanceDiagnostics,
-                            onRowAction: handleSettingsRowAction
-                        )
-                        .frame(maxWidth: ShellSettingsMetrics.contentWidth, alignment: .leading)
-                        .padding(.leading, ShellSettingsMetrics.detailContentLeadingPadding)
-                        .padding(.trailing, ShellSettingsMetrics.detailContentTrailingPadding)
-                        .padding(.top, ShellSettingsMetrics.detailContentTopPadding)
-                        .padding(.bottom, ShellSettingsMetrics.detailContentBottomPadding)
-                        .frame(maxWidth: .infinity, alignment: .topLeading)
-                    }
-                    .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
-                }
-                .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
-            }
-            .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
-        }
-        .task(id: refreshTaskID) {
-            await refreshSettingsSummaries()
-        }
-        .sheet(isPresented: $isManagedUserCreationPresented) {
-            ShellManagedUserCreationSheet(
-                draft: $managedUserCreationDraft,
-                previewResult: managedUserCreationPreviewResult,
-                diagnostics: managedUserApplyDiagnostics,
-                isApplying: managedUserApplyInFlight,
-                onDraftChanged: resetManagedUserCreationPreview,
-                onPreview: reviewManagedUserCreationDraft,
-                onApply: applyManagedUserCreationPreview,
-                onCancel: {
-                    isManagedUserCreationPresented = false
-                }
-            )
-        }
-        .sheet(item: $managedUserActionSheet) { sheet in
-            ShellManagedUserPlanSheet(
-                sheet: sheet,
-                diagnostics: managedUserApplyDiagnostics,
-                isApplying: managedUserApplyInFlight,
-                onApply: {
-                    applyManagedUserActionSheet(sheet)
-                },
-                onCancel: {
-                    managedUserActionSheet = nil
-                }
-            )
-        }
-    }
-
-    private var refreshTaskID: String {
-        descriptor.contentID ?? descriptor.title
-    }
-
-    private var sidebarVisible: Binding<Bool> {
-        Binding(
-            get: { !isSidebarCollapsed },
-            set: { isSidebarCollapsed = !$0 }
-        )
-    }
-
-    private var performanceDiagnosticsBinding: Binding<Bool> {
-        Binding(
-            get: { performanceDiagnosticsEnabled },
-            set: { nextValue in
-                performanceDiagnosticsEnabled = nextValue
-                AlanPerformanceDiagnosticsController.shared.setEnabled(nextValue)
-            }
-        )
-    }
-
-    private var diagnosticsSummary: ShellSettingsDiagnosticsSummary {
-        let summary = AlanPerformanceDiagnosticsController.shared.summarySnapshot()
-        return ShellSettingsDiagnosticsSummary(
-            isEnabled: performanceDiagnosticsEnabled,
-            retainedEventCount: AlanPerformanceDiagnosticsController.shared.eventsSnapshot().count,
-            stutterMarkerCount: summary.stutterMarkerCount,
-            lastExportURL: lastDiagnosticsExportURL
-        )
-    }
-
-    private func exportPerformanceDiagnostics() {
-        lastDiagnosticsExportURL = AlanPerformanceDiagnosticsExportPresenter.exportRecentDiagnostics(
-            installChannel: localSummary.channelLabel
-        )
-    }
-
-    @discardableResult
-    @MainActor
-    private func refreshLocalTerminalIdentitySummaries() -> ManagedTerminalAccountSettingsSummary {
-        let profiles = TerminalProfileSettingsSummary.current()
-        let accounts = ManagedTerminalAccountSettingsSummary.current(
-            terminalProfiles: profiles,
-            helperClient: AlanPrivilegedHelperAppClient(channel: .current())
-        )
-        terminalProfilesSummary = profiles
-        managedTerminalAccountsSummary = accounts
-        return accounts
-    }
-
-    @MainActor
-    private func handleSettingsRowAction(
-        row: ShellSettingsRowModel,
-        action: ShellSettingsRowActionKind
-    ) {
-        managedUserApplyDiagnostics = []
-        switch action {
-        case .create:
-            managedUserCreationDraft = ManagedTerminalUserCreationDraft(
-                unixUserName: "",
-                displayLabel: ""
-            )
-            managedUserCreationPreviewResult = nil
-            isManagedUserCreationPresented = true
-        case .review, .repair, .verify, .remove:
-            handleExistingManagedUserAction(row: row, action: action)
-        case .installHelper, .updateHelper, .uninstallHelper:
-            applyPrivilegedHelperLifecycleAction(action)
-        }
-    }
-
-    @MainActor
-    private func applyPrivilegedHelperLifecycleAction(_ action: ShellSettingsRowActionKind) {
-        let manager = AlanPrivilegedHelperAppServiceManager()
-        let result: AlanPrivilegedHelperLifecycleResult
-        switch action {
-        case .installHelper, .updateHelper:
-            result = manager.installOrUpdate()
-        case .uninstallHelper:
-            result = manager.uninstall()
-        case .create, .review, .repair, .verify, .remove:
-            return
-        }
-        privilegedHelperSummary = PrivilegedHelperSettingsSummary(status: result.status)
-        managedUserApplyDiagnostics = result.diagnostic.map {
-            [$0.sanitizedMessage, "Credentials redacted."]
-        } ?? ["Privileged helper \(result.action.rawValue) completed. Credentials redacted."]
-    }
-
-    @MainActor
-    private func handleExistingManagedUserAction(
-        row: ShellSettingsRowModel,
-        action: ShellSettingsRowActionKind
-    ) {
-        guard let plan = managedUserPlan(forRowID: row.id) else { return }
-        switch action {
-        case .create, .installHelper, .updateHelper, .uninstallHelper:
-            return
-        case .review:
-            managedUserActionSheet = ShellManagedUserActionSheetState(
-                action: action,
-                plan: plan,
-                allowsApply: false
-            )
-        case .repair:
-            managedUserActionSheet = ShellManagedUserActionSheetState(
-                action: action,
-                plan: plan,
-                allowsApply: true
-            )
-        case .verify:
-            let refreshedSummary = refreshLocalTerminalIdentitySummaries()
-            let refreshedPlan = refreshedSummary.plans.first {
-                $0.request.accountName == plan.request.accountName
-            } ?? plan
-            managedUserActionSheet = ShellManagedUserActionSheetState(
-                action: .review,
-                plan: refreshedPlan,
-                allowsApply: false
-            )
-        case .remove:
-            let helperClient = AlanPrivilegedHelperAppClient(channel: .current())
-            let status = helperClient.status()
-            let diagnosis = status.isHealthy
-                ? helperClient.diagnoseManagedUser(plan.request)
-                : AlanManagedUserDiagnosis.helperUnavailable(request: plan.request, status: status)
-            let rollbackPlan = ManagedTerminalAccountPlanner.rollbackPlan(
-                request: plan.request,
-                diagnosis: diagnosis,
-                scope: .alanIntegrationOnly,
-                terminalProfiles: terminalProfilesSummary.document
-            )
-            managedUserActionSheet = ShellManagedUserActionSheetState(
-                action: action,
-                plan: rollbackPlan,
-                allowsApply: true
-            )
-        }
-    }
-
-    private func managedUserPlan(forRowID rowID: String) -> ManagedTerminalAccountPlan? {
-        let prefix = "terminalAccount."
-        guard rowID.hasPrefix(prefix) else { return nil }
-        let accountName = String(rowID.dropFirst(prefix.count))
-        return managedTerminalAccountsSummary.plans.first {
-            $0.request.accountName == accountName
-        }
-    }
-
-    @MainActor
-    private func resetManagedUserCreationPreview() {
-        managedUserCreationPreviewResult = nil
-        managedUserApplyDiagnostics = []
-    }
-
-    @MainActor
-    private func reviewManagedUserCreationDraft() {
-        let request = managedUserCreationDraft.request
-        let helperClient = AlanPrivilegedHelperAppClient(channel: .current())
-        let status = helperClient.status()
-        let diagnosis = status.isHealthy
-            ? helperClient.diagnoseManagedUser(request)
-            : AlanManagedUserDiagnosis.helperUnavailable(request: request, status: status)
-        managedUserCreationPreviewResult = ManagedTerminalUserCreationPreviewBuilder.make(
-            draft: managedUserCreationDraft,
-            existingUsers: managedTerminalAccountsSummary.users,
-            terminalProfiles: terminalProfilesSummary,
-            diagnosis: diagnosis
-        )
-    }
-
-    @MainActor
-    private func applyManagedUserCreationPreview(_ preview: ManagedTerminalUserCreationPreview) {
-        guard !managedUserApplyInFlight else { return }
-        managedUserApplyInFlight = true
-        managedUserApplyDiagnostics = ["Applying managed user changes. Credentials redacted."]
-
-        Task {
-            defer { managedUserApplyInFlight = false }
-            let result = await Self.applyManagedUserPlanInBackground(
-                plan: preview.plan,
-                request: preview.request
-            )
-            if Task.isCancelled { return }
-
-            terminalProfilesSummary = result.terminalProfiles
-            managedTerminalAccountsSummary = result.managedTerminalAccounts
-            managedUserApplyDiagnostics = result.applyResult.visibleDiagnostics
-            if !result.applyResult.cancelled && result.applyResult.failedStep == nil {
-                isManagedUserCreationPresented = false
-                managedUserCreationPreviewResult = nil
-            }
-        }
-    }
-
-    @MainActor
-    private func applyManagedUserActionSheet(_ sheet: ShellManagedUserActionSheetState) {
-        guard !managedUserApplyInFlight else { return }
-        managedUserApplyInFlight = true
-        managedUserApplyDiagnostics = ["Applying managed user changes. Credentials redacted."]
-
-        Task {
-            defer { managedUserApplyInFlight = false }
-            let result = await Self.applyManagedUserPlanInBackground(
-                plan: sheet.plan,
-                request: sheet.plan.request
-            )
-            if Task.isCancelled { return }
-
-            terminalProfilesSummary = result.terminalProfiles
-            managedTerminalAccountsSummary = result.managedTerminalAccounts
-            managedUserApplyDiagnostics = result.applyResult.visibleDiagnostics
-            if !result.applyResult.cancelled && result.applyResult.failedStep == nil {
-                managedUserActionSheet = nil
-            }
-        }
-    }
-
-    nonisolated private static func applyManagedUserPlanInBackground(
-        plan: ManagedTerminalAccountPlan,
-        request: ManagedTerminalAccountRequest
-    ) async -> ShellManagedUserApplyBackgroundResult {
-        await withCheckedContinuation { continuation in
-            let continuationBox = ShellManagedUserApplyContinuationBox(continuation: continuation)
-            let work = Task.detached(priority: .userInitiated) {
-                let result = runManagedUserPlanInBackground(plan: plan, request: request)
-                continuationBox.resume(returning: result)
-            }
-            Task.detached(priority: .userInitiated) {
-                try? await Task.sleep(nanoseconds: managedUserApplyTimeoutNanoseconds)
-                guard !work.isCancelled else { return }
-                work.cancel()
-                managedUserApplyLogger.error("Managed User apply timed out.")
-                continuationBox.resume(returning: timeoutManagedUserApplyResult(plan: plan))
-            }
-        }
-    }
-
-    nonisolated private static func runManagedUserPlanInBackground(
-        plan: ManagedTerminalAccountPlan,
-        request: ManagedTerminalAccountRequest
-    ) -> ShellManagedUserApplyBackgroundResult {
-        managedUserApplyLogger.info("Managed User apply started.")
-        let catalogStore = ManagedTerminalAccountCatalogStore.defaultStore()
-        let isRemovalPlan = plan.steps.contains {
-            switch $0.kind {
-            case .removeManagedTerminalProfile, .deleteAccount, .deleteHomeDirectory:
-                return true
-            case .helperStep(let helperKind):
-                return helperKind == .removeManagedUserIntegration
-                    || helperKind == .deleteAccount
-                    || helperKind == .deleteHomeDirectory
-            default:
-                return false
-            }
-        } && !plan.steps.contains {
-            switch $0.kind {
-            case .createStandardAccount,
-                 .repairAccountType,
-                 .repairHomeDirectory,
-                 .repairShell,
-                 .hideAccount,
-                 .createOrUpdateTerminalProfile:
-                return true
-            case .helperStep(let helperKind):
-                switch helperKind {
-                case .removeManagedUserIntegration, .deleteAccount, .deleteHomeDirectory:
-                    return false
-                case .createStandardAccount,
-                     .repairAccountType,
-                     .repairHomeDirectory,
-                     .repairShell,
-                     .hideAccount,
-                     .writeOwnershipMarker,
-                     .verifyAccount,
-                     .verifyManagedUserPTY:
-                    return true
-                }
-            case .removeManagedTerminalProfile, .deleteAccount, .deleteHomeDirectory:
-                return false
-            }
-        }
-        if !isRemovalPlan {
-            try? catalogStore.upsert(
-                ManagedTerminalAccountCatalogEntry(
-                    accountName: request.accountName,
-                    displayLabel: request.fullName ?? request.accountName
-                )
-            )
-        }
-        let channel = AlanInstallChannel.current()
-        let helperClient = AlanPrivilegedHelperAppClient(channel: channel)
-        let executor = ManagedTerminalAccountHelperExecutor(
-            channel: channel,
-            helperClient: helperClient
-        )
-        let applyResult = executor.apply(plan)
-        if isRemovalPlan && !applyResult.cancelled && applyResult.failedStep == nil {
-            try? catalogStore.remove(accountName: request.accountName)
-        }
-        let terminalProfiles = TerminalProfileSettingsSummary.current()
-        let managedTerminalAccounts = ManagedTerminalAccountSettingsSummary.current(
-            terminalProfiles: terminalProfiles,
-            helperClient: helperClient
-        )
-        managedUserApplyLogger.info("Managed User apply finished.")
-        return ShellManagedUserApplyBackgroundResult(
-            applyResult: applyResult,
-            terminalProfiles: terminalProfiles,
-            managedTerminalAccounts: managedTerminalAccounts
-        )
-    }
-
-    nonisolated private static func timeoutManagedUserApplyResult(
-        plan: ManagedTerminalAccountPlan
-    ) -> ShellManagedUserApplyBackgroundResult {
-        let terminalProfiles = TerminalProfileSettingsSummary.current()
-        return ShellManagedUserApplyBackgroundResult(
-            applyResult: ManagedTerminalAccountApplyResult(
-                completedSteps: [],
-                failedStep: plan.steps.first?.kind,
-                cancelled: false,
-                visibleDiagnostics: [
-                    "Managed User apply timed out. Credentials redacted.",
-                ]
-            ),
-            terminalProfiles: terminalProfiles,
-            managedTerminalAccounts: ManagedTerminalAccountSettingsSummary.current(
-                terminalProfiles: terminalProfiles,
-                helperClient: AlanPrivilegedHelperAppClient(channel: .current())
-            )
-        )
-    }
-
-    private struct ShellManagedUserApplyBackgroundResult {
-        let applyResult: ManagedTerminalAccountApplyResult
-        let terminalProfiles: TerminalProfileSettingsSummary
-        let managedTerminalAccounts: ManagedTerminalAccountSettingsSummary
-    }
-
-    private final class ShellManagedUserApplyContinuationBox: @unchecked Sendable {
-        private let lock = NSLock()
-        private var hasResumed = false
-        private let continuation: CheckedContinuation<ShellManagedUserApplyBackgroundResult, Never>
-
-        init(continuation: CheckedContinuation<ShellManagedUserApplyBackgroundResult, Never>) {
-            self.continuation = continuation
-        }
-
-        func resume(returning result: ShellManagedUserApplyBackgroundResult) {
-            lock.lock()
-            defer { lock.unlock() }
-            guard !hasResumed else { return }
-            hasResumed = true
-            continuation.resume(returning: result)
-        }
-    }
-
-    @MainActor
-    private func refreshSettingsSummaries() async {
-        localSummary = ShellSettingsLocalSummary.current()
-        privilegedHelperSummary = PrivilegedHelperSettingsSummary.current()
-        refreshLocalTerminalIdentitySummaries()
-    }
-}
-
-private struct ShellSettingsBackdrop: View {
-    var body: some View {
-        ZStack {
-            ShellPalette.settingsPane
-            ShellPalette.windowBackdropTint.opacity(0.025)
-        }
-    }
-}
-
-private struct ShellSettingsNavigationRailBackground: View {
-    var body: some View {
-        Color.clear
-    }
-}
-
-private struct ShellSettingsDetailBackground: View {
-    var body: some View {
-        Color.clear
-    }
-}
-
-private struct ShellSettingsNavigationView: View {
-    @Environment(\.accessibilityReduceMotion) private var reduceMotion
-    @State private var hoveredGroup: ShellSettingsNavigationGroup?
-
-    let groups: [ShellSettingsNavigationGroupModel]
-    @Binding var selectedGroup: ShellSettingsNavigationGroup
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: ShellSettingsMetrics.navigationRowSpacing) {
-            ForEach(groups) { group in
-                Button {
-                    selectedGroup = group.id
-                } label: {
-                    HStack(spacing: ShellSettingsMetrics.navigationRowContentSpacing) {
-                        Image(systemName: group.systemName)
-                            .font(ShellSettingsTypography.navigationIcon)
-                            .foregroundStyle(iconStyle(for: group))
-                            .frame(width: ShellSettingsMetrics.navigationIconSlotWidth, height: 16)
-
-                        Text(group.title)
-                            .font(ShellSettingsTypography.navigationLabel(selected: group.id == selectedGroup))
-                            .foregroundStyle(textStyle(for: group))
-                            .lineLimit(1)
-
-                        Spacer(minLength: 0)
-                    }
-                    .padding(.horizontal, ShellSettingsMetrics.navigationRowHorizontalPadding)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                    .frame(height: ShellSettingsMetrics.navigationRowHeight)
-                    .contentShape(Rectangle())
-                    .background {
-                        ShellSettingsNavigationRowBackground(
-                            state: rowVisualState(for: group)
-                        )
-                    }
-                }
-                .buttonStyle(.plain)
-                .help(group.title)
-                .accessibilityLabel(Text(group.title))
-                .onHover { isHovered in
-                    if isHovered {
-                        hoveredGroup = group.id
-                    } else if hoveredGroup == group.id {
-                        hoveredGroup = nil
-                    }
-                }
-            }
-        }
-        .animation(reduceMotion ? nil : .easeOut(duration: 0.12), value: hoveredGroup)
-        .animation(reduceMotion ? nil : .easeOut(duration: 0.14), value: selectedGroup)
-    }
-
-    private func iconStyle(for group: ShellSettingsNavigationGroupModel) -> some ShapeStyle {
-        group.id == selectedGroup
-            ? AnyShapeStyle(ShellPalette.settingsPrimaryInk)
-            : AnyShapeStyle(ShellPalette.settingsSecondaryInk)
-    }
-
-    private func textStyle(for group: ShellSettingsNavigationGroupModel) -> some ShapeStyle {
-        group.id == selectedGroup
-            ? AnyShapeStyle(ShellPalette.settingsPrimaryInk)
-            : AnyShapeStyle(ShellPalette.settingsSecondaryInk)
-    }
-
-    private func rowVisualState(
-        for group: ShellSettingsNavigationGroupModel
-    ) -> ShellSettingsNavigationRowVisualState {
-        if group.id == selectedGroup {
-            return .selected
-        }
-
-        if group.id == hoveredGroup {
-            return .hover
-        }
-
-        return .normal
-    }
-}
-
-private enum ShellSettingsNavigationRowVisualState: Equatable {
-    case normal
-    case hover
-    case selected
-
-    var fill: Color? {
-        switch self {
-        case .normal:
-            return nil
-        case .hover:
-            return ShellPalette.settingsNavigationHover
-        case .selected:
-            return ShellPalette.settingsNavigationSelection
-        }
-    }
-
-    var stroke: Color {
-        switch self {
-        case .normal:
-            return .clear
-        case .hover:
-            return ShellPalette.line.opacity(0.07)
-        case .selected:
-            return ShellPalette.line.opacity(0.08)
-        }
-    }
-}
-
-private struct ShellSettingsNavigationRowBackground: View {
-    let state: ShellSettingsNavigationRowVisualState
-
-    var body: some View {
-        let shape = RoundedRectangle(
-            cornerRadius: ShellSettingsMetrics.navigationSelectionCornerRadius,
-            style: .continuous
-        )
-
-        ZStack(alignment: .leading) {
-            if let fill = state.fill {
-                shape
-                    .fill(fill)
-                    .overlay {
-                        shape.stroke(state.stroke, lineWidth: 0.5)
-                    }
-            }
-        }
-    }
-}
-
-private struct ShellSettingsGroupView: View {
-    let group: ShellSettingsNavigationGroupModel
-    @Binding var appearanceMode: ShellAppearanceMode
-    let sidebarVisible: Binding<Bool>
-    @Binding var dimsInactiveSplitPanes: Bool
-    let performanceDiagnosticsEnabled: Binding<Bool>
-    let onExportPerformanceDiagnostics: () -> Void
-    let onRowAction: (ShellSettingsRowModel, ShellSettingsRowActionKind) -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: ShellSettingsMetrics.pageTitleToSectionsSpacing) {
-            Text(group.title)
-                .font(ShellSettingsTypography.pageTitle)
-                .foregroundStyle(ShellPalette.settingsPrimaryInk)
-
-            VStack(alignment: .leading, spacing: ShellSettingsMetrics.sectionSpacing) {
-                ForEach(group.sections) { section in
-                    ShellSettingsSectionView(
-                        section: section,
-                        appearanceMode: $appearanceMode,
-                        sidebarVisible: sidebarVisible,
-                        dimsInactiveSplitPanes: $dimsInactiveSplitPanes,
-                        performanceDiagnosticsEnabled: performanceDiagnosticsEnabled,
-                        onExportPerformanceDiagnostics: onExportPerformanceDiagnostics,
-                        onRowAction: onRowAction
-                    )
-                }
-            }
-        }
-    }
-}
-
-private struct ShellSettingsVerticalDivider: View {
-    var body: some View {
-        Rectangle()
-            .fill(ShellPalette.line.opacity(0.13))
-            .frame(width: 0.8)
-    }
-}
-
-private struct ShellSettingsSectionView: View {
-    let section: ShellSettingsGroupSectionModel
-    @Binding var appearanceMode: ShellAppearanceMode
-    let sidebarVisible: Binding<Bool>
-    @Binding var dimsInactiveSplitPanes: Bool
-    let performanceDiagnosticsEnabled: Binding<Bool>
-    let onExportPerformanceDiagnostics: () -> Void
-    let onRowAction: (ShellSettingsRowModel, ShellSettingsRowActionKind) -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            HStack(alignment: .firstTextBaseline, spacing: 14) {
-                Text(section.title.uppercased())
-                    .font(ShellSettingsTypography.sectionTitle)
-                    .foregroundStyle(ShellPalette.settingsSecondaryInk)
-                    .tracking(0.4)
-                    .lineLimit(1)
-
-                Rectangle()
-                    .fill(ShellPalette.line.opacity(0.20))
-                    .frame(height: 0.8)
-            }
-            .padding(.bottom, ShellSettingsMetrics.sectionTitleBottomPadding)
-
-            VStack(spacing: 0) {
-                ForEach(Array(section.rows.enumerated()), id: \.element.id) { index, row in
-                    if index > 0 {
-                        ShellSettingsDivider()
-                    }
-
-                    rowView(row)
-                }
-            }
-        }
-    }
-
-    @ViewBuilder
-    private func rowView(_ row: ShellSettingsRowModel) -> some View {
-        switch row.id {
-        case "appearance":
-            ShellSettingsRow(
-                systemName: row.systemName,
-                title: row.title,
-                detail: row.detail
-            ) {
-                Picker("Appearance", selection: $appearanceMode) {
-                    ForEach(ShellAppearanceMode.allCases) { mode in
-                        Text(mode.label).tag(mode)
-                    }
-                }
-                .labelsHidden()
-                .pickerStyle(.segmented)
-                .controlSize(.small)
-                .frame(width: ShellSettingsMetrics.segmentedControlWidth)
-            }
-        case "sidebar":
-            ShellSettingsRow(
-                systemName: row.systemName,
-                title: row.title,
-                detail: row.detail
-            ) {
-                Toggle(row.title, isOn: sidebarVisible)
-                    .labelsHidden()
-                    .toggleStyle(.switch)
-                    .controlSize(.small)
-            }
-        case "inactiveSplitDimming":
-            ShellSettingsRow(
-                systemName: row.systemName,
-                title: row.title,
-                detail: row.detail
-            ) {
-                Toggle(row.title, isOn: $dimsInactiveSplitPanes)
-                    .labelsHidden()
-                    .toggleStyle(.switch)
-                    .controlSize(.small)
-            }
-        case "performanceDiagnostics":
-            ShellSettingsRow(
-                systemName: row.systemName,
-                title: row.title,
-                detail: row.detail
-            ) {
-                Toggle(row.title, isOn: performanceDiagnosticsEnabled)
-                    .labelsHidden()
-                    .toggleStyle(.switch)
-                    .controlSize(.small)
-            }
-        case "performanceDiagnosticsExport":
-            ShellSettingsRow(
-                systemName: row.systemName,
-                title: row.title,
-                detail: row.detail
-            ) {
-                Button("Export", action: onExportPerformanceDiagnostics)
-                    .buttonStyle(.bordered)
-                    .controlSize(.small)
-                    .disabled(!performanceDiagnosticsEnabled.wrappedValue)
-                    .opacity(
-                        performanceDiagnosticsEnabled.wrappedValue
-                            ? 1
-                            : ShellSettingsMetrics.disabledButtonOpacity
-                    )
-            }
-        case "applicationSupport", "dataRoot":
-            ShellSettingsRow(
-                systemName: row.systemName,
-                title: row.title,
-                detail: row.value
-            ) {
-                ShellSettingsPathAction(value: row.value)
-            }
-        default:
-            if !row.actions.isEmpty {
-                ShellSettingsRow(
-                    systemName: row.systemName,
-                    title: row.title,
-                    detail: row.detail
-                ) {
-                    ShellSettingsRowActionAccessory(row: row, onAction: onRowAction)
-                }
-            } else if let detail = row.detail, row.value != nil {
-                ShellSettingsRow(
-                    systemName: row.systemName,
-                    title: row.title,
-                    detail: detail
-                ) {
-                    ShellSettingsValueLabel(
-                        value: row.value,
-                        mutability: row.mutability
-                    )
-                }
-            } else {
-                ShellSettingsRow(
-                    systemName: row.systemName,
-                    title: row.title,
-                    detail: row.detail ?? row.value
-                )
-            }
-        }
-    }
-}
-
-private struct ShellSettingsRowHoveredKey: EnvironmentKey {
-    static let defaultValue = false
-}
-
-private extension EnvironmentValues {
-    var shellSettingsRowHovered: Bool {
-        get { self[ShellSettingsRowHoveredKey.self] }
-        set { self[ShellSettingsRowHoveredKey.self] = newValue }
-    }
-}
-
-private struct ShellSettingsRow<Accessory: View>: View {
-    @State private var isHovered = false
-
-    let systemName: String
-    let title: String
-    let detail: String?
-    @ViewBuilder let accessory: () -> Accessory
-
-    init(
-        systemName: String,
-        title: String,
-        detail: String?,
-        @ViewBuilder accessory: @escaping () -> Accessory
-    ) {
-        self.systemName = systemName
-        self.title = title
-        self.detail = detail
-        self.accessory = accessory
-    }
-
-    var body: some View {
-        HStack(alignment: .center, spacing: ShellSettingsMetrics.rowColumnSpacing) {
-            VStack(alignment: .leading, spacing: ShellSettingsMetrics.rowTextSpacing) {
-                Text(title)
-                    .font(ShellSettingsTypography.rowTitle)
-                    .foregroundStyle(ShellPalette.settingsPrimaryInk)
-                    .lineLimit(1)
-
-                if let detail,
-                   !detail.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
-                {
-                    Text(detail)
-                        .font(ShellSettingsTypography.rowDetail)
-                        .foregroundStyle(ShellPalette.settingsSecondaryInk)
-                        .lineSpacing(1)
-                        .lineLimit(2)
-                        .fixedSize(horizontal: false, vertical: true)
-                }
-            }
-            .frame(maxWidth: .infinity, alignment: .leading)
-            .layoutPriority(1)
-
-            accessoryView
-                .environment(\.shellSettingsRowHovered, isHovered)
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .padding(.vertical, ShellSettingsMetrics.rowVerticalPadding)
-        .frame(minHeight: rowMinHeight)
-        .onHover { isHovered = $0 }
-    }
-
-    private var rowMinHeight: CGFloat {
-        let hasDetail = detail?.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty == false
-        return hasDetail ? ShellSettingsMetrics.rowMinHeightWithDetail : ShellSettingsMetrics.rowMinHeight
-    }
-
-    @ViewBuilder
-    private var accessoryView: some View {
-        accessory()
-            .font(ShellSettingsTypography.accessory)
-            .frame(width: ShellSettingsMetrics.accessoryColumnWidth, alignment: .trailing)
-    }
-}
-
-private extension ShellSettingsRow where Accessory == EmptyView {
-    init(
-        systemName: String,
-        title: String,
-        detail: String?
-    ) {
-        self.systemName = systemName
-        self.title = title
-        self.detail = detail
-        self.accessory = { EmptyView() }
-    }
-}
-
-private struct ShellSettingsValueLabel: View {
-    let value: String?
-    let mutability: ShellSettingsRowMutability
-
-    var body: some View {
-        HStack(spacing: 6) {
-            Text(value ?? "Unavailable")
-                .font(ShellSettingsTypography.value)
-                .foregroundStyle(valueStyle)
-                .lineLimit(1)
-                .truncationMode(.middle)
-                .multilineTextAlignment(.trailing)
-                .fixedSize(horizontal: false, vertical: true)
-        }
-        .frame(maxWidth: ShellSettingsMetrics.valueColumnWidth, alignment: .trailing)
-        .help(value ?? "Unavailable")
-    }
-
-    private var valueStyle: some ShapeStyle {
-        if value == "Unavailable" {
-            return AnyShapeStyle(ShellPalette.settingsDisabledInk)
-        }
-        if mutability == .actionOnly {
-            return AnyShapeStyle(ShellPalette.settingsValueInk)
-        }
-        return AnyShapeStyle(ShellPalette.settingsValueInk)
-    }
-}
-
-private struct ShellSettingsInlineValueAction: View {
-    @Environment(\.shellSettingsRowHovered) private var isRowHovered
-
-    let isEnabled: Bool
-    let buttonSystemName: String
-    let buttonHelp: String
-    let action: () -> Void
-
-    var body: some View {
-        Button(action: action) {
-            Label("Copy", systemImage: buttonSystemName)
-                .labelStyle(.titleAndIcon)
-        }
-        .buttonStyle(.bordered)
-        .controlSize(.small)
-        .disabled(!isEnabled)
-        .opacity(isRowHovered || isEnabled ? 1 : ShellSettingsMetrics.disabledButtonOpacity)
-        .help(buttonHelp)
-    }
-}
-
-private struct ShellSettingsPathAction: View {
-    let value: String?
-
-    var body: some View {
-        Button("Show…") {
-            shellSettingsOpenFolder(value)
-        }
-        .buttonStyle(.bordered)
-        .controlSize(.small)
-        .disabled(!ShellLocalFolderOpener.canOpenFolder(displayPath: value))
-        .help(value ?? "Folder unavailable")
-    }
-}
-
-private struct ShellSettingsRowActionAccessory: View {
-    let row: ShellSettingsRowModel
-    let onAction: (ShellSettingsRowModel, ShellSettingsRowActionKind) -> Void
-
-    var body: some View {
-        HStack(spacing: ShellSettingsMetrics.inlineActionSpacing) {
-            if let value = row.value,
-               !value.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
-                Text(value)
-                    .font(ShellSettingsTypography.value)
-                    .foregroundStyle(ShellPalette.settingsValueInk)
-                    .lineLimit(1)
-                    .truncationMode(.tail)
-                    .frame(maxWidth: 96, alignment: .trailing)
-            }
-
-            if row.actions.count == 1,
-               let action = row.actions.first {
-                Button {
-                    onAction(row, action.id)
-                } label: {
-                    Label(action.title, systemImage: action.systemName)
-                        .labelStyle(.iconOnly)
-                }
-                .buttonStyle(.bordered)
-                .controlSize(.small)
-                .help(action.title)
-            } else {
-                Menu {
-                    ForEach(row.actions) { action in
-                        Button {
-                            onAction(row, action.id)
-                        } label: {
-                            Label(action.title, systemImage: action.systemName)
-                        }
-                    }
-                } label: {
-                    Label("Actions", systemImage: "ellipsis.circle")
-                        .labelStyle(.iconOnly)
-                }
-                .menuStyle(.button)
-                .buttonStyle(.bordered)
-                .controlSize(.small)
-                .help("Actions")
-            }
-        }
-        .frame(maxWidth: .infinity, alignment: .trailing)
-    }
-}
-
-private struct ShellManagedUserActionSheetState: Identifiable, Equatable {
-    let action: ShellSettingsRowActionKind
-    let plan: ManagedTerminalAccountPlan
-    let allowsApply: Bool
-
-    var id: String {
-        "\(action.rawValue)-\(plan.request.accountName)-\(plan.steps.map(\.kind).count)"
-    }
-
-    var title: String {
-        switch action {
-        case .create:
-            return "Create Managed User"
-        case .review:
-            return "Review Managed User"
-        case .repair:
-            return "Repair Managed User"
-        case .verify:
-            return "Verify Managed User"
-        case .remove:
-            return "Remove Managed User"
-        case .installHelper, .updateHelper, .uninstallHelper:
-            return "Managed User Helper"
-        }
-    }
-
-    var applyTitle: String {
-        switch action {
-        case .remove:
-            return "Remove"
-        case .repair:
-            return "Apply Repair"
-        case .create:
-            return "Create"
-        case .review, .verify:
-            return "Apply"
-        case .installHelper, .updateHelper, .uninstallHelper:
-            return "Apply"
-        }
-    }
-}
-
-private struct ShellManagedUserCreationSheet: View {
-    @Binding var draft: ManagedTerminalUserCreationDraft
-    let previewResult: ManagedTerminalUserCreationPreviewResult?
-    let diagnostics: [String]
-    let isApplying: Bool
-    let onDraftChanged: () -> Void
-    let onPreview: () -> Void
-    let onApply: (ManagedTerminalUserCreationPreview) -> Void
-    let onCancel: () -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            Text("Create Managed User")
-                .font(ShellSettingsTypography.pageTitle)
-                .foregroundStyle(ShellPalette.settingsPrimaryInk)
-
-            VStack(alignment: .leading, spacing: 10) {
-                TextField("Unix user", text: $draft.unixUserName)
-                TextField("Display label", text: $draft.displayLabel)
-            }
-            .textFieldStyle(.roundedBorder)
-            .disabled(isApplying)
-
-            previewContent
-
-            HStack {
-                Button("Cancel", role: .cancel, action: onCancel)
-                    .disabled(isApplying)
-                Spacer()
-                if isApplying {
-                    ProgressView()
-                        .controlSize(.small)
-                }
-                Button("Review Plan", action: onPreview)
-                    .disabled(isApplying)
-                Button("Apply") {
-                    if let preview = previewResult?.preview {
-                        onApply(preview)
-                    }
-                }
-                .disabled(isApplying || (previewResult?.preview?.plan.steps.isEmpty ?? true))
-            }
-        }
-        .padding(24)
-        .frame(width: 460, alignment: .leading)
-        .onChange(of: draft) { _, _ in
-            onDraftChanged()
-        }
-    }
-
-    @ViewBuilder
-    private var previewContent: some View {
-        if let result = previewResult,
-           !result.errors.isEmpty {
-            VStack(alignment: .leading, spacing: 6) {
-                ForEach(result.errors.map(errorMessage), id: \.self) { message in
-                    Label(message, systemImage: "exclamationmark.triangle")
-                        .font(ShellSettingsTypography.rowDetail)
-                        .foregroundStyle(ShellPalette.settingsSecondaryInk)
-                }
-            }
-        } else if let preview = previewResult?.preview {
-            VStack(alignment: .leading, spacing: 6) {
-                ForEach(preview.visiblePlanRows, id: \.self) { row in
-                    Label(row, systemImage: "checkmark")
-                        .font(ShellSettingsTypography.rowDetail)
-                        .foregroundStyle(ShellPalette.settingsSecondaryInk)
-                }
-            }
-        }
-
-        if !diagnostics.isEmpty {
-            VStack(alignment: .leading, spacing: 4) {
-                ForEach(diagnostics, id: \.self) { diagnostic in
-                    Text(diagnostic)
-                        .font(ShellSettingsTypography.rowDetail)
-                        .foregroundStyle(ShellPalette.settingsSecondaryInk)
-                }
-            }
-        }
-    }
-
-    private func errorMessage(_ error: ManagedTerminalUserCreationPreviewError) -> String {
-        switch error {
-        case .missingUnixUserName:
-            return "Unix user is required."
-        case .missingDisplayLabel:
-            return "Display label is required."
-        case .duplicateUnixUser(let user):
-            return "\(user) already exists."
-        case .terminalProfileConflict(let profileID):
-            return "Terminal Profile \(profileID) already exists."
-        case .validation:
-            return "Use a valid local Unix user name."
-        }
-    }
-}
-
-private struct ShellManagedUserPlanSheet: View {
-    let sheet: ShellManagedUserActionSheetState
-    let diagnostics: [String]
-    let isApplying: Bool
-    let onApply: () -> Void
-    let onCancel: () -> Void
-
-    private var preview: ManagedTerminalUserCreationPreview {
-        ManagedTerminalUserCreationPreview(request: sheet.plan.request, plan: sheet.plan)
-    }
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            Text(sheet.title)
-                .font(ShellSettingsTypography.pageTitle)
-                .foregroundStyle(ShellPalette.settingsPrimaryInk)
-
-            Text("\(sheet.plan.request.fullName ?? sheet.plan.request.accountName) · \(planStatusText)")
-                .font(ShellSettingsTypography.rowDetail)
-                .foregroundStyle(ShellPalette.settingsSecondaryInk)
-
-            VStack(alignment: .leading, spacing: 6) {
-                ForEach(preview.visiblePlanRows, id: \.self) { row in
-                    Label(row, systemImage: "checkmark")
-                        .font(ShellSettingsTypography.rowDetail)
-                        .foregroundStyle(ShellPalette.settingsSecondaryInk)
-                }
-            }
-
-            if !diagnostics.isEmpty {
-                VStack(alignment: .leading, spacing: 4) {
-                    ForEach(diagnostics, id: \.self) { diagnostic in
-                        Text(diagnostic)
-                            .font(ShellSettingsTypography.rowDetail)
-                            .foregroundStyle(ShellPalette.settingsSecondaryInk)
-                    }
-                }
-            }
-
-            HStack {
-                Button("Close", role: .cancel, action: onCancel)
-                    .disabled(isApplying)
-                Spacer()
-                if isApplying {
-                    ProgressView()
-                        .controlSize(.small)
-                }
-                if sheet.allowsApply {
-                    Button(sheet.applyTitle, action: onApply)
-                        .disabled(isApplying || sheet.plan.steps.isEmpty)
-                }
-            }
-        }
-        .padding(24)
-        .frame(width: 460, alignment: .leading)
-    }
-
-    private var planStatusText: String {
-        switch sheet.plan.status {
-        case .alreadyReady:
-            return "Ready"
-        case .readyToApply:
-            return "Ready to apply"
-        case .repair:
-            return "Repairable"
-        case .helperUnavailable:
-            return "Helper unavailable"
-        case .accountNotAlanManaged:
-            return "Not managed"
-        case .ptySpawnFailed:
-            return "PTY failed"
-        case .invalid:
-            return "Invalid"
-        case .requiresDestructiveConfirmation:
-            return "Needs confirmation"
-        case .terminalProfileConflict:
-            return "Terminal Profile conflict"
-        }
-    }
-}
-
-@MainActor
-private func shellSettingsCopyToPasteboard(_ value: String?) {
-    guard let value = value?.trimmingCharacters(in: .whitespacesAndNewlines),
-          !value.isEmpty
-    else {
-        return
-    }
-    ShellSystemPasteboard().writeString(value)
-}
-
-@MainActor
-private func shellSettingsOpenFolder(_ value: String?) {
-    ShellLocalFolderOpener.openFolder(displayPath: value)
-}
-
-private struct ShellSettingsDivider: View {
-    var body: some View {
-        Divider()
-            .opacity(0.45)
-            .padding(.leading, ShellSettingsMetrics.rowDividerLeadingPadding)
-    }
-}
-
-private enum ShellSettingsMetrics {
-    static let navigationWidth: CGFloat = 188
-    static let navigationLeadingPadding: CGFloat = 12
-    static let navigationTrailingPadding: CGFloat = 8
-    static let navigationTopPadding: CGFloat = 24
-    static let navigationRowHeight: CGFloat = 30
-    static let navigationRowSpacing: CGFloat = 2
-    static let navigationRowHorizontalPadding: CGFloat = 8
-    static let navigationRowContentSpacing: CGFloat = 12
-    static let navigationIconSlotWidth: CGFloat = 18
-    static let navigationSelectionCornerRadius: CGFloat = 7
-    static let contentWidth: CGFloat = 760
-    static let detailContentLeadingPadding: CGFloat = 48
-    static let detailContentTrailingPadding: CGFloat = 48
-    static let detailContentTopPadding: CGFloat = 42
-    static let detailContentBottomPadding: CGFloat = 40
-    static let pageTitleToSectionsSpacing: CGFloat = 26
-    static let sectionSpacing: CGFloat = 28
-    static let sectionTitleBottomPadding: CGFloat = 10
-    static let rowVerticalPadding: CGFloat = 8
-    static let rowMinHeight: CGFloat = 48
-    static let rowMinHeightWithDetail: CGFloat = 56
-    static let agentSummaryRowMinHeight: CGFloat = 58
-    static let rowTextSpacing: CGFloat = 1
-    static let rowColumnSpacing: CGFloat = 20
-    static let rowDividerLeadingPadding: CGFloat = 0
-    static let accessoryColumnWidth: CGFloat = 188
-    static let valueColumnWidth: CGFloat = 220
-    static let inlineActionSpacing: CGFloat = 8
-    static let inlineIconButtonSize: CGFloat = 22
-    static let segmentedControlWidth: CGFloat = 196
-    static let disabledButtonOpacity: CGFloat = 0.55
-}
-
-private enum ShellSettingsTypography {
-    static let navigationIcon = Font.system(size: 13, weight: .regular)
-
-    static func navigationLabel(selected: Bool) -> Font {
-        .system(size: 13, weight: selected ? .semibold : .regular)
-    }
-
-    static let pageTitle = Font.system(size: 22, weight: .semibold)
-    static let sectionTitle = Font.system(size: 11, weight: .semibold)
-    static let rowTitle = Font.system(size: 13, weight: .semibold)
-    static let rowDetail = Font.system(size: 12, weight: .regular)
-    static let accessory = Font.system(size: 13, weight: .regular)
-    static let value = Font.system(size: 13, weight: .medium)
-    static let agentName = Font.system(size: 15, weight: .semibold)
-    static let badge = Font.system(size: 11.5, weight: .medium)
-    static let valueActionIcon = Font.system(size: 9.5, weight: .semibold)
-    static let inlineActionIcon = Font.system(size: 10.5, weight: .medium)
-    static let actionButton = Font.system(size: 12.3, weight: .medium)
 }
 
 private struct ShellContentPaneTitleBarView: View {
