+++ clients/apple/scripts/test-shell-settings-surface.sh
@@ -12,7 +12,11 @@ mkdir -p "$MODULE_CACHE_DIR"
 
 cargo build -p alan-shell-core-ffi
 
-TERMINAL_PANE_VIEW="$REPO_ROOT/clients/apple/alan-macos/TerminalPaneView.swift"
+SHELL_SETTINGS_VIEW_FILES=(
+    "$REPO_ROOT/clients/apple/alan-macos/Views/Shell/Settings/ShellSettingsContentView.swift"
+    "$REPO_ROOT/clients/apple/alan-macos/Views/Shell/Settings/ShellSettingsNavigationView.swift"
+    "$REPO_ROOT/clients/apple/alan-macos/Views/Shell/Settings/ShellSettingsComponents.swift"
+)
 SHELL_SETTINGS_MODEL_FILES=(
     "$REPO_ROOT/clients/apple/alan-macos/Models/Shell/ShellSettingsSurfaceModel.swift"
     "$REPO_ROOT/clients/apple/alan-macos/Models/Shell/TerminalSettingsSummaries.swift"
@@ -21,31 +25,33 @@ SHELL_SETTINGS_MODEL_FILES=(
     "$REPO_ROOT/clients/apple/alan-macos/Models/Shell/ManagedTerminalUserSettings.swift"
     "$REPO_ROOT/clients/apple/alan-macos/Models/Shell/ShellSettingsHostSummaries.swift"
 )
-if grep -q 'TextField("Mac user"' "$TERMINAL_PANE_VIEW"; then
+if grep -q 'TextField("Mac user"' "${SHELL_SETTINGS_VIEW_FILES[@]}"; then
     echo "Managed User creation form must not expose Mac user as a primary input." >&2
     exit 1
 fi
-if grep -q 'ManagedTerminalUserProvisioningFlow.applyApproved' "$TERMINAL_PANE_VIEW"; then
+if grep -q 'ManagedTerminalUserProvisioningFlow.applyApproved' "${SHELL_SETTINGS_VIEW_FILES[@]}"; then
     echo "Managed User Settings actions must not run privileged apply synchronously on the UI path." >&2
     exit 1
 fi
-if grep -q 'guard !Task.isCancelled else { return }' "$TERMINAL_PANE_VIEW"; then
+if grep -q 'guard !Task.isCancelled else { return }' "${SHELL_SETTINGS_VIEW_FILES[@]}"; then
     echo "Managed User apply tasks must clear applying state even if the Swift task is cancelled." >&2
     exit 1
 fi
-if ! grep -q 'defer { managedUserApplyInFlight = false }' "$TERMINAL_PANE_VIEW"; then
+if ! grep -q 'defer { managedUserApplyInFlight = false }' "${SHELL_SETTINGS_VIEW_FILES[@]}"; then
     echo "Managed User apply tasks must use deferred UI cleanup so spinners cannot leak." >&2
     exit 1
 fi
-if ! grep -q 'Managed User apply timed out' "$TERMINAL_PANE_VIEW"; then
+if ! grep -q 'Managed User apply timed out' "${SHELL_SETTINGS_VIEW_FILES[@]}"; then
     echo "Managed User apply tasks must surface a timeout instead of spinning indefinitely." >&2
     exit 1
 fi
-if grep -Fq 'managedUserApplyTimeoutNanoseconds: UInt64 = 90 * 1_000_000_000' "$TERMINAL_PANE_VIEW"; then
+if grep -Fq 'managedUserApplyTimeoutNanoseconds: UInt64 = 90 * 1_000_000_000' \
+    "${SHELL_SETTINGS_VIEW_FILES[@]}"
+then
     echo "Managed User apply timeout must allow enough time for macOS administrator approval." >&2
     exit 1
 fi
-if ! grep -Fq '10 * 60 * 1_000_000_000' "$TERMINAL_PANE_VIEW"; then
+if ! grep -Fq '10 * 60 * 1_000_000_000' "${SHELL_SETTINGS_VIEW_FILES[@]}"; then
     echo "Managed User apply timeout must be a documented 10 minute administrator approval budget." >&2
     exit 1
 fi
